refactored UpdateCalcFromInput.js
renamed all occurances of UpdateCalcFromInput.js to UpdateCalcVar
renamed UpdateCalcFromInput.js to UpdateCalcVar.js
updated docstring
